Sprint 1

+++ src/main/java/com/ust/pos/user/service/impl/UserServiceImpl.java
@@ -7,6 +7,8 @@ import com.ust.pos.user.service.UserService;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +19,7 @@ import java.util.Optional;
 @Service
 public class UserServiceImpl implements UserService {
 
+    public static final String USER_WITH_USERNAME_EMAIL = "User with username/email - ";
     @Autowired
     private UserRepository userRepository;
 
@@ -36,7 +39,7 @@ public class UserServiceImpl implements UserService {
         String username = userDto.getUsername();
         User existingUser = userRepository.findByUsername(username);
         if (existingUser != null) {
-            userDto.setMessage("User with username/email - " + userDto.getUsername() + " already exists");
+            userDto.setMessage(USER_WITH_USERNAME_EMAIL + userDto.getUsername() + " already exists");
             userDto.setSuccess(false);
             return userDto;
         }
@@ -52,17 +55,18 @@ public class UserServiceImpl implements UserService {
         Optional<User> userOptional = userRepository.findById(userDto.getId());
 
         if (userOptional.isEmpty()) {
-            userDto.setMessage("User with username/email - " + userDto.getUsername() + " not found");
+            userDto.setMessage(USER_WITH_USERNAME_EMAIL +
+                    userDto.getUsername() + " not found");
             userDto.setSuccess(false);
             return userDto;
         } else {
             User existingUser = userOptional.get();
-            if (!username.equalsIgnoreCase(existingUser.getUsername())) {
-                if (userRepository.findByUsername(username) != null) {
-                    userDto.setMessage("User with username/email - " + userDto.getUsername() + " already exists");
+            if (!username.equalsIgnoreCase(existingUser.getUsername())&&(userRepository.findByUsername(username) != null)){
+                    userDto.setMessage(USER_WITH_USERNAME_EMAIL
+                            + userDto.getUsername() + " already exists");
                     userDto.setSuccess(false);
                     return userDto;
-                }
+
             }
             modelMapper.map(userDto, existingUser);
             userRepository.save(existingUser);
@@ -71,10 +75,26 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean delete(String username) {
-        return userRepository.deleteByUsername(username);
-    }
 
+    public UserDto delete(String username) {
+
+        Authentication authentication =
+                SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            String loggedInUsername = authentication.getName();
+            User response = userRepository.findByUsername(username);
+            UserDto user = modelMapper.map(response, UserDto.class);
+            if (username.equals(loggedInUsername)) {
+                user.setMessage("Cannot delete the logged in User");
+                user.setSuccess(false);
+                return user;
+            }
+
+            userRepository.deleteByUsername(username);
+            return user;
+        }
+        return  null;
+    }
     @Override
     public List<UserDto> findAll() {
         Type listType = new TypeToken<List<UserDto>>() {
